% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgexp.R
\docType{data}
\name{imgexp}
\alias{imgexp}
\title{A bitmap digital image}
\format{
numeric dataset
}
\usage{
imgexp
}
\description{
The sample dataset is a bitmap digital image from a published Kaplan Meier curves. \cr
            It is the same image we used to extract the sample \code{\link{Radiationdata}}.
            We can use \code{getpoints()} function to extract the coordinates from the image.
}
\examples{

str(imgexp)
plot.new()
rasterImage(imgexp, 0, 0, 1, 1)

}
\references{
Bonner JA, Harari PM, Giralt J, Azarnia N, Shin DM, Cohen RB, Jones CU, Sur R,
Raben D, Jassem J, Ove R, Kies MS, Baselga J, Youssoufian H, Amellal N, Rowinsky EK, Ang KK:
Radiotherapy plus Cetuximab for Squamous-Cell Carcinoma of the Head and Neck. N Engl J Med.
2006, 354: 567-78. 10.1056/NEJMoa053422.
}
\keyword{datasets}
