% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreport.R
\name{survreport}
\alias{survreport}
\title{Survival analysis on the reconstructed individual patients data (IPD).}
\usage{
survreport(ipd1,ipd2=NULL,arms=1,interval=6,s=c(0.75,0.5,0.25))
}
\arguments{
\item{ipd1}{The first individual patients' lifetable with three columns (time,status, and the treatment indicators), found in the output of \code{\link{getIPD}} function.  \cr}

\item{ipd2}{The second indivdual patients' lifetable with three columns (time,status, and the treatment indicators), found in the output of \code{\link{getIPD}} function.  \cr}

\item{arms}{This indicate the functional analysis of data related to either one or two treatments: can be either 1 or 2.}

\item{interval}{The time intervals where we want to report the survival rates: the default is at every 6 months time interval.}

\item{s}{The survival rates want to report the corresponding times}
}
\description{
This function will graph the Kaplan-Meier curves and the cumulative hazard curves from the
            reconstructed IPD (the output of \code{\link{getIPD}} function). It will also report
            the times when survival rates reach certain percentages, and the survival rates at some
            certain time points (for example, at every half-year time points).
}
\examples{
### Get data from the sample dataset=======================
radio <- Radiationdata$radio
radioplus <- Radiationdata$radioplus
trisk <- Radiationdata$trisk
nrisk_radio <- Radiationdata$nrisk.radio
nrisk_radioplus <- Radiationdata$nrisk.radioplus
### Estimate the IPD for the Radiotherapy treatment group ====================
pre_radio <- preprocess(dat=radio, trisk=trisk,nrisk=nrisk_radio,maxy=100)
est_radio <- getIPD(prep=pre_radio,armind=0,tot.events=NULL)
### Estimate the IPD for the Radiotherapy plus treatment group ====================
pre_radio_plus <- preprocess(dat=radioplus, trisk=trisk,nrisk=nrisk_radioplus,maxy=100)
est_radio_plus <- getIPD(prep=pre_radio_plus,armind=1,tot.events=NULL)
### survival report for one arm ===================
survreport(ipd1=est_radio$IPD,arms=1,interval=6,s=c(0.8,0.5,0.3))
survreport(ipd1=est_radio_plus$IPD,arms=1,interval=10)
### survival report for two arms ===================
survreport(ipd1=est_radio$IPD,ipd2=est_radio_plus$IPD,arms=2,interval=8)


}
\references{
Guyot P, Ades AE, Ouwens MJ, Welton NJ. Enhanced secondary analysis of survival data: reconstructing the data from published Kaplan-Meier survival curves. BMC Med Res Methodol.2012; 1:9.
}
